Making tests easier to read and write.
Creates a new set of Ops: TOp, and T2Op, that can inspect or alter a plan.
This allows creating unit tests that are just a call to applyAllOpsToPlan.

Also add applyAllOpsToPlanAndThenInverse since calling the inverse on a stack
of Ops is also common in tests.

The differnece between TOp and T2Op is that the callback function
is not called on inverse in TOp, but is called during inverse for T2Op.