Add Span and use in Slack.
Create Span, which encapsulates the start and finish times of a task, and then use
Spans in Slack. Also just use Span in renderer so we can swap from plotting
early to late w/o a change to the rendering code.

Also consolidate another set of calls into a single call to edgesBySrcAndDstToMap.